+++ apps/api/src/yolink/yolink.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { YolinkHome } from './entities/yolink-home.entity';
+import { YolinkDevice } from './entities/yolink-device.entity';
 import { YolinkService } from './yolink.service';
 import { YolinkController } from './yolink.controller';
 import { AlertsModule } from '../alerts/alerts.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([YolinkHome]), AlertsModule],
+  imports: [TypeOrmModule.forFeature([YolinkHome, YolinkDevice]), AlertsModule],
   controllers: [YolinkController],
   providers: [YolinkService],
   exports: [YolinkService],